cron

+++ sql/MPR/mpr.sql
@@ -58,9 +58,9 @@ MPR as (
 select
 	Date, PlatformId, SoftwareName , Gateway, Vertical,ParentAccountId, ParentName , 
 	FeePaymentType, PaymentTypeGroup ,
-	sum(TPV_USD)::money as TPV_USD, (sum(TPV_USD) - sum(TPV_Net_USD))::money as "Refunds/Chargebacks", sum(TPV_Net_USD)::money as TPV_Net_USD ,
-	sum(Txn_Count)::int as Txn_Count, 	
-	sum(Revenue_Net_USD)::money as Revenue_Net_USD, sum(COGS_USD)::money as COGS_USD
+	sum(TPV_USD) as TPV_USD, (sum(TPV_USD) - sum(TPV_Net_USD)) as "Refunds/Chargebacks", sum(TPV_Net_USD) as TPV_Net_USD ,
+	sum(Txn_Count) as Txn_Count, 	
+	sum(Revenue_Net_USD) as Revenue_Net_USD, sum(COGS_USD) as COGS_USD
 from
 	MPR
 group by
